Fix controlled checkbox issue
React does not pass properties to a component when they are undefined.

When reading from local storage we may get an `undefined`. That means that
using the value from local storage as the `checked` value of the
checkbox may cause it to become "uncontrolled" and then controlled again
once we do get a value from local storage.